packages refactored

+++ src/main/java/bullbear/app/dto/auth/BlockchainTx.java
@@ -3,13 +3,15 @@ package bullbear.app.dto.auth;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 @AllArgsConstructor
 public class BlockchainTx {
     private String hash;           // tx_hash
     private String from;
     private String to;
-    private Double amount;         // in USDT
+    private BigDecimal amount;         // in USDT
     private int confirmations;
     private String token;          // USDT
 }
